+++ src/main/java/controller/LottoController.java
@@ -1,0 +1,5 @@
+package controller;
+
+public interface LottoController {
+    void run();
+}
