--- src/main/java/lotto_controller/LottoController.java
@@ -1,5 +1,0 @@
-package lotto_controller;
-
-public interface LottoController {
-    void run();
-}
