--- src/main/java/lotto_model/comparator/LottoComparator.java
@@ -1,7 +1,0 @@
-package lotto_model.comparator;
-
-import lotto_model.lotto.Lotto;
-
-public interface LottoComparator {
-    void compare(Lotto tragetLotto, int bonusNumber);
-}
