--- src/main/java/lotto_model/data/LottoRepository.java
@@ -1,10 +1,0 @@
-package lotto_model.data;
-
-import lotto_model.lotto.Lotto;
-
-import java.util.List;
-
-public interface LottoRepository {
-    List<Lotto> getLottoSaved();
-    void save(Lotto lotto);
-}
