--- src/main/java/lotto_model/generator/LottoGenerator.java
@@ -1,5 +1,0 @@
-package lotto_model.generator;
-
-public interface LottoGenerator {
-    void printLottoSaved(int inputMoney);
-}
