--- src/main/java/lotto_model/service/LottoService.java
@@ -1,8 +1,0 @@
-package lotto_model.service;
-
-import lotto_model.lotto.Lotto;
-
-public interface LottoService {
-    void lottoCompare(Lotto targetLotto, int bonusNumber);
-    void printLotto(int inputMoney);
-}
