--- src/main/java/lotto_view/LottoInput.java
@@ -1,9 +1,0 @@
-package lotto_view;
-
-import lotto_model.lotto.Lotto;
-
-public interface LottoInput {
-    int getBonusNumber();
-    int getInputMoney();
-    Lotto inputNumber();
-}
