--- src/main/java/lotto_view/LottoOutput.java
@@ -1,5 +1,0 @@
-package lotto_view;
-
-public interface LottoOutput {
-    void resultPrint(int inputMoney);
-}
