+++ src/main/java/model/comparator/LottoComparator.java
@@ -1,0 +1,7 @@
+package model.comparator;
+
+import model.lotto.Lotto;
+
+public interface LottoComparator {
+    void compare(Lotto tragetLotto, int bonusNumber);
+}
