+++ src/main/java/model/data/LottoRepository.java
@@ -1,0 +1,10 @@
+package model.data;
+
+import model.lotto.Lotto;
+
+import java.util.List;
+
+public interface LottoRepository {
+    List<Lotto> getLottoSaved();
+    void save(Lotto lotto);
+}
