+++ src/main/java/model/data/LottoWinning.java
@@ -1,0 +1,13 @@
+package model.data;
+
+import model.lotto.LottoRank;
+
+public interface LottoWinning {
+    void addFirst();
+    void addSecond();
+    void addThird();
+    void addFourth();
+    void addFifth();
+
+    int getQuantity(LottoRank lottoRank);
+}
