+++ src/main/java/model/generator/LottoGenerator.java
@@ -1,0 +1,5 @@
+package model.generator;
+
+public interface LottoGenerator {
+    void printLottoSaved(int inputMoney);
+}
