+++ src/main/java/model/lotto/Lotto.java
@@ -1,0 +1,28 @@
+package model.lotto;
+
+import java.util.*;
+
+import static exception.ErrorMessage.*;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    public boolean contains(Integer element) {
+        return numbers.contains(element);
+    }
+
+    public List<Integer> getNumbers() {
+        return numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException(INVALID_NUMBER_COUNT.getErrorMessage());
+        }
+    }
+}
