+++ src/main/java/model/service/LottoService.java
@@ -1,0 +1,8 @@
+package model.service;
+
+import model.lotto.Lotto;
+
+public interface LottoService {
+    void lottoCompare(Lotto targetLotto, int bonusNumber);
+    void printLotto(int inputMoney);
+}
