+++ src/main/java/view/LottoInput.java
@@ -1,0 +1,9 @@
+package view;
+
+import model.lotto.Lotto;
+
+public interface LottoInput {
+    int getBonusNumber();
+    int getInputMoney();
+    Lotto inputNumber();
+}
