+++ src/main/java/view/LottoOutput.java
@@ -1,0 +1,5 @@
+package view;
+
+public interface LottoOutput {
+    void resultPrint(int inputMoney);
+}
